# openapi.yml
openapi: 3.0.0
info:
  title: FoodSea Unified API
  description: Unified REST API for core, optimization, and ordering services
    behind ingress.
  version: 1.0.0
paths:
...
  /auth/login:
    post:
      description: Authenticates a user and returns JWT tokens
      tags:
        - auth
      summary: Login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/internal_modules_identity_handler.LoginRequest"
        description: Login credentials
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/github_com_foodsea_core_internal_platform_httputil.\
                      Response"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/internal_modules_identity_handler.AuthResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_foodsea_core_internal_platform_httputil.\
                  Response"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_foodsea_core_internal_platform_httputil.\
                  Response"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_foodsea_core_internal_platform_httputil.\
                  Response"
  /auth/logout:
    post:
      security:
        - BearerAuth: []
      description: Revokes all refresh tokens for the current user
      tags:
        - auth
      summary: Logout
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_foodsea_core_internal_platform_httputil.\
                  Response"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_foodsea_core_internal_platform_httputil.\
                  Response"
  /auth/refresh:
    post:
      description: Exchanges a valid refresh token for a new token pair
      tags:
        - auth
      summary: Refresh tokens
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/internal_modules_identity_handler.RefreshRequest"
        description: Refresh token
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/github_com_foodsea_core_internal_platform_httputil.\
                      Response"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/internal_modules_identity_handler.TokenPairResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_foodsea_core_internal_platform_httputil.\
                  Response"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_foodsea_core_internal_platform_httputil.\
                  Response"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_foodsea_core_internal_platform_httputil.\
                  Response"
  /auth/register:
    post:
      description: Creates a new user account with phone or email and returns JWT tokens
      tags:
        - auth
      summary: Register a new user
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/internal_modules_identity_handler.RegisterRequest"
        description: Registration credentials
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/github_com_foodsea_core_internal_platform_httputil.\
                      Response"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/internal_modules_identity_handler.AuthResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_foodsea_core_internal_platform_httputil.\
                  Response"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_foodsea_core_internal_platform_httputil.\
                  Response"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_foodsea_core_internal_platform_httputil.\
                  Response"
...
servers:
  - url: http://dev.111.88.159.219.nip.io/api/v1
components:
  securitySchemes:
    BearerAuth:
      description: 'JWT токен. Формат: "Bearer {token}"'
      type: apiKey
      name: Authorization
      in: header
  schemas:
...
    internal_modules_identity_handler.AuthResponse:
      type: object
      properties:
        access_expires_at:
          type: string
        access_token:
          type: string
        refresh_expires_at:
          type: string
        refresh_token:
          type: string
        user:
          $ref: "#/components/schemas/internal_modules_identity_handler.UserResponse"
    internal_modules_identity_handler.LoginRequest:
      type: object
      required:
        - password
      properties:
        email:
          type: string
        password:
          type: string
        phone:
          type: string
    internal_modules_identity_handler.RefreshRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
    internal_modules_identity_handler.RegisterRequest:
      type: object
      required:
        - password
      properties:
        email:
          type: string
        password:
          type: string
          minLength: 8
        phone:
          type: string
    internal_modules_identity_handler.TokenPairResponse:
      type: object
      properties:
        access_expires_at:
          type: string
        access_token:
          type: string
        refresh_expires_at:
          type: string
        refresh_token:
          type: string
    internal_modules_identity_handler.UserResponse:
      type: object
      properties:
        created_at:
          type: string
        email:
          type: string
        id:
          type: string
        onboarding_done:
          type: boolean
        phone:
          type: string
        updated_at:
          type: string
...